events: support 32 bit system IDs
libevents held system IDs in 8 bit fields, so passing the now wider ones
in truncated them, which MSVC flags as C4267 on 32 bit Windows where
size_t and uint32_t are the same type.

Point libevents at mavlink/libevents#pr-sysid32, which widens them, and
follow the one API change that came with it: REQUEST_EVENT's target does
not fit the struct handed to the send callback, so the callback now gets
it separately. Pack the fields rather than encoding the struct so that a
target above 255 ends up in the extended header.

Note that an event still cannot address a system above 255, since EVENT's
destination_system is not a routing target and so stays 8 bit. Such a
system only sees broadcast events until the message itself grows a wider
field.

The libevents tag needs to move to a merged commit before this lands.

+++ cpp/src/mavsdk/plugins/events/event_handler.hpp
@@ -25,7 +25,7 @@ public:
         handle_event_f handle_event_cb,
         health_and_arming_checks_updated_f health_and_arming_checks_updated_cb,
         SystemImpl& system_impl,
-        uint8_t system_id,
+        uint32_t system_id,
         uint8_t component_id);
     ~EventHandler();
